+++ packages/orq-rc/src/models/operations/createdatasetitem.ts
@@ -2355,7 +2355,7 @@ export const Evaluations3$inboundSchema: z.ZodType<
   ),
   reviewed_by_id: z.string(),
   reviewed_at: z.string().datetime({ offset: true }).default(
-    "2026-06-09T13:49:51.201Z",
+    "2026-06-09T18:01:25.899Z",
   ).transform(v => new Date(v)),
   type: z.literal("string_array"),
   values: z.array(z.string()),
@@ -2400,7 +2400,7 @@ export const Evaluations2$inboundSchema: z.ZodType<
   source: CreateDatasetItemEvaluationsSource$inboundSchema.default("orq"),
   reviewed_by_id: z.string(),
   reviewed_at: z.string().datetime({ offset: true }).default(
-    "2026-06-09T13:49:51.200Z",
+    "2026-06-09T18:01:25.898Z",
   ).transform(v => new Date(v)),
   type: z.literal("number"),
   value: z.number(),
@@ -2446,7 +2446,7 @@ export const Evaluations1$inboundSchema: z.ZodType<
   source: EvaluationsSource$inboundSchema.default("orq"),
   reviewed_by_id: z.string(),
   reviewed_at: z.string().datetime({ offset: true }).default(
-    "2026-06-09T13:49:51.199Z",
+    "2026-06-09T18:01:25.897Z",
   ).transform(v => new Date(v)),
   type: z.literal("string"),
   value: z.string(),
@@ -2529,7 +2529,7 @@ export const CreateDatasetItemResponseBody$inboundSchema: z.ZodType<
   created: z.string().datetime({ offset: true }).transform(v => new Date(v))
     .optional(),
   updated: z.string().datetime({ offset: true }).default(
-    "2026-06-09T13:49:27.367Z",
+    "2026-06-09T18:01:04.261Z",
   ).transform(v => new Date(v)),
 }).transform((v) => {
   return remap$(v, {